using 'long long' for pointer address

+++ clang/test/Analysis/pointer-sub.c
@@ -12,9 +12,9 @@ void f1(void) {
   d = &x - (&x + 1); // no-warning
   d = (&x + 0) - &x; // no-warning
   d = (z + 10) - z; // no-warning
-  d = (unsigned long)&y - (unsigned long)&x; // no-warning
-  unsigned long l = 1;
-  d = l - (unsigned long)&y; // no-warning
+  d = (long long)&y - (long long)&x; // no-warning
+  long long l = 1;
+  d = l - (long long)&y; // no-warning
   Ptr p1 = &x;
   Ptr p2 = &y;
   d = p1 - p2; // expected-warning{{Subtraction of two pointers that do not point into the same array is undefined behavior}}
@@ -37,8 +37,8 @@ void f2(void) {
   d = (int *)((char *)(&a[4]) + sizeof(int)) - &a[4]; // no-warning (pointers into the same array data)
   d = (int *)((char *)(&a[4]) + 1) - &a[4]; // expected-warning{{Subtraction of two pointers that}}
 
-  long a1 = (long)&a[1];
-  long b1 = (long)&b[1];
+  long long a1 = (long long)&a[1];
+  long long b1 = (long long)&b[1];
   d = a1 - b1;
 }
